fix: resolving use of StState Context and AutoSizer.
