+++ nerdlets/neon-nerdlet/board-admin.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Icon, Tabs, TabsItem, TextField, Button } from 'nr1';
+import { Tabs, TabsItem, TextField, Button } from 'nr1';
 
 export default class BoardAdmin extends React.Component {
   static propTypes = {
