+++ nerdlets/neon-nerdlet/board.js
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import {
-  nerdlet,
   AccountStorageQuery,
   AccountStorageMutation,
   NerdGraphQuery,
